fix(frontend): split skill-hub/[name] into server/client components for static export
